refactor: move query related files to /query
