+++ frontend/src/modules/organizations/organizations-table/table.tsx
@@ -9,8 +9,8 @@ import { showToast } from '~/lib/toasts';
 import ContentPlaceholder from '~/modules/common/content-placeholder';
 import { DataTable } from '~/modules/common/data-table';
 import type { OrganizationsSearch } from '~/modules/organizations/organizations-table';
-import { organizationsQueryOptions } from '~/modules/organizations/organizations-table/helpers/query-options';
 import { useDataFromSuspenseInfiniteQuery } from '~/query/hooks/use-data-from-query';
+import { organizationsQueryOptions } from '~/query/options/infinite-query-options';
 import { useUserStore } from '~/store/user';
 import type { BaseTableMethods, BaseTableProps, Organization } from '~/types/common';
