+++ frontend/src/modules/users/user-sheet.tsx
@@ -1,18 +1,11 @@
-import { queryOptions, useQuery } from '@tanstack/react-query';
+import { useQuery } from '@tanstack/react-query';
 import { FlameKindling, WifiOff } from 'lucide-react';
 import { useTranslation } from 'react-i18next';
-import { getUser } from '~/api/users';
 import { useOnlineManager } from '~/hooks/use-online-manager';
 import ContentPlaceholder from '~/modules/common/content-placeholder';
 import Spinner from '~/modules/common/spinner';
 import UserProfilePage from '~/modules/users/profile-page';
-import { usersKeys } from '~/query/query-key-factories';
-
-export const userQueryOptions = (idOrSlug: string) =>
-  queryOptions({
-    queryKey: usersKeys.single(idOrSlug),
-    queryFn: () => getUser(idOrSlug),
-  });
+import { userQueryOptions } from '~/query/options/query-options';
 
 const UserSheet = ({ idOrSlug, orgIdOrSlug }: { idOrSlug: string; orgIdOrSlug?: string }) => {
   const { t } = useTranslation();
