+++ frontend/src/modules/users/users-table/table.tsx
@@ -8,9 +8,9 @@ import { useMutation } from '~/hooks/use-mutations';
 import { showToast } from '~/lib/toasts';
 import { DataTable } from '~/modules/common/data-table';
 import type { UsersSearch } from '~/modules/users/users-table';
-import { usersQueryOptions } from '~/modules/users/users-table/helpers/query-options';
 import { useDataFromSuspenseInfiniteQuery } from '~/query/hooks/use-data-from-query';
 import { useMutateQueryData } from '~/query/hooks/use-mutate-query-data';
+import { usersQueryOptions } from '~/query/options/infinite-query-options';
 import { usersKeys } from '~/query/query-key-factories';
 import type { BaseTableMethods, BaseTableProps, User } from '~/types/common';
