+++ frontend/src/offline-config.tsx
@@ -1,8 +1,6 @@
-import { organizationQueryOptions } from '~/modules/organizations/organization-page';
+import { attachmentsQueryOptions, membersQueryOptions } from '~/query/options/infinite-query-options';
+import { organizationQueryOptions } from '~/query/options/query-options';
 import type { UserMenuItem } from '~/types/common';
-
-import { attachmentsQueryOptions } from '~/modules/attachments/attachments-table/helpers/query-options';
-import { membersQueryOptions } from '~/modules/organizations/members-table/helpers/query-options';
 
 // This function returns queries that need to be prefetched based on the entity of the item.
 // It is used to prefetch data for each unarchived item in user menu if offlineAccess is enabled, allowing the app to have necessary data while offline.
