+++ frontend/src/query/index.tsx
@@ -2,9 +2,9 @@ import { QueryClientProvider as BaseQueryClientProvider } from '@tanstack/react-
 import { PersistQueryClientProvider } from '@tanstack/react-query-persist-client';
 import { useEffect } from 'react';
 import { persister, queryClient } from '~/lib/router';
-import { meQueryOptions, menuQueryOptions } from '~/modules/users/helpers/query-options';
 import { queriesToMap } from '~/offline-config';
 import { prefetchQuery, waitFor } from '~/query/helpers';
+import { meQueryOptions, menuQueryOptions } from '~/query/options/query-options';
 import { useGeneralStore } from '~/store/general';
 import type { UserMenuItem } from '~/types/common';
