+++ frontend/src/query/mutations/members.ts
@@ -41,7 +41,7 @@ const onError = (_: Error, __: UpdateMembershipProp & RemoveMembersProps, contex
 queryClient.setMutationDefaults(membersKeys.update(), {
   mutationFn: updateMembership,
   onSuccess: async (updatedMembership, { idOrSlug, entityType, orgIdOrSlug }) => {
-    const exactKey = membersKeys.list({ idOrSlug, entityType, orgIdOrSlug });
+    const exactKey = membersKeys.table({ idOrSlug, entityType, orgIdOrSlug });
     const similarKey = membersKeys.similar({ idOrSlug, entityType, orgIdOrSlug });
 
     const queries = getQueries<Member>(exactKey, similarKey);
@@ -71,7 +71,7 @@ queryClient.setMutationDefaults(membersKeys.delete(), {
     const { ids, idOrSlug, entityType, orgIdOrSlug } = variables;
 
     const context: MemberContextProp[] = [];
-    const exactKey = membersKeys.list({ idOrSlug, entityType, orgIdOrSlug });
+    const exactKey = membersKeys.table({ idOrSlug, entityType, orgIdOrSlug });
     const similarKey = membersKeys.similar({ idOrSlug, entityType, orgIdOrSlug });
 
     const queries = await getCancelingRefetchQueries<Member>(exactKey, similarKey);
