+++ frontend/src/query/query-key-factories.ts
@@ -16,7 +16,8 @@ export const attachmentsKeys = {
 
 export const membersKeys = {
   all: ['members'] as const,
-  list: (filters?: GetMembersParams) => [...membersKeys.all, 'list', filters] as const,
+  list: () => [...membersKeys.all, 'list'] as const,
+  table: (filters?: GetMembersParams) => [...membersKeys.list(), filters] as const,
   similar: (filters?: Pick<GetMembersParams, 'orgIdOrSlug' | 'idOrSlug' | 'entityType'>) => [...membersKeys.all, filters] as const,
   update: () => [...membersKeys.all, 'update'] as const,
   delete: () => [...membersKeys.all, 'delete'] as const,
