+++ frontend/src/routes/organizations.tsx
@@ -4,11 +4,10 @@ import { Suspense, lazy } from 'react';
 import { z } from 'zod';
 import { offlineFetch, offlineFetchInfinite } from '~/lib/query-client';
 import { queryClient } from '~/lib/router';
-import { attachmentsQueryOptions } from '~/modules/attachments/attachments-table/helpers/query-options';
 import ErrorNotice from '~/modules/common/error-notice';
-import { membersQueryOptions } from '~/modules/organizations/members-table/helpers/query-options';
-import { organizationQueryOptions } from '~/modules/organizations/organization-page';
 import { baseEntityRoutes } from '~/nav-config';
+import { attachmentsQueryOptions, membersQueryOptions } from '~/query/options/infinite-query-options';
+import { organizationQueryOptions } from '~/query/options/query-options';
 import { AppRoute } from '~/routes/general';
 import type { Organization as OrganizationType } from '~/types/common';
 import { noDirectAccess } from '~/utils/no-direct-access';
